+++ apps/api/src/auth/types.ts
@@ -1,0 +1,45 @@
+import {
+  PublicKeyCredentialCreationOptionsJSON,
+  Base64URLString,
+  CredentialDeviceType,
+  AuthenticatorTransportFuture,
+} from '@simplewebauthn/types';
+
+export type UserModel = {
+  id: string;
+  username: string;
+  passkeyOptions: PublicKeyCredentialCreationOptionsJSON;
+};
+
+/**
+ * It is strongly advised that credentials get their own DB
+ * table, ideally with a foreign key somewhere connecting it
+ * to a specific UserModel.
+ *
+ * "SQL" tags below are suggestions for column data types and
+ * how best to store data received during registration for use
+ * in subsequent authentications.
+ */
+export type Passkey = {
+  // SQL: Store as `TEXT`. Index this column
+  id: Base64URLString;
+  // SQL: Store raw bytes as `BYTEA`/`BLOB`/etc...
+  //      Caution: Node ORM's may map this to a Buffer on retrieval,
+  //      convert to Uint8Array as necessary
+  publicKey: Uint8Array;
+  // SQL: Foreign Key to an instance of your internal user model
+  user: UserModel;
+  // SQL: Store as `TEXT`. Index this column. A UNIQUE constraint on
+  //      (webAuthnUserID + user) also achieves maximum user privacy
+  webAuthnUserID: Base64URLString;
+  // SQL: Consider `BIGINT` since some authenticators return atomic timestamps as counters
+  counter: number;
+  // SQL: `VARCHAR(32)` or similar, longest possible value is currently 12 characters
+  // Ex: 'singleDevice' | 'multiDevice'
+  deviceType: CredentialDeviceType;
+  // SQL: `BOOL` or whatever similar type is supported
+  backedUp: boolean;
+  // SQL: `VARCHAR(255)` and store string array as a CSV string
+  // Ex: ['ble' | 'cable' | 'hybrid' | 'internal' | 'nfc' | 'smart-card' | 'usb']
+  transports?: AuthenticatorTransportFuture[];
+};
